# backend/api.yaml
openapi: 3.0.0
info:
  title: HMS API – Hospital Management System
  version: 1.0.0
  description: >
    REST API for HMS (Hospital Management System) built for MAD-2.
    Supports JWT authentication, Admin/Doctor/Patient roles, Doctor management,
    Appointment booking, and Treatment recording.

servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        specialization: { type: string }

    DoctorInput:
      type: object
      required: [username, name, email, specialization, password]
      properties:
        username: { type: string }
        name: { type: string }
        email: { type: string }
        specialization: { type: string }
        password: { type: string }

    RegisterInput:
      type: object
      required: [username, name, email, password]
      properties:
        username: { type: string }
        name: { type: string }
        email: { type: string }
        password: { type: string }

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    LoginOutput:
      type: object
      properties:
        token: { type: string }
        role: { type: string }

    AppointmentInput:
      type: object
      required: [doctor_id, date, time]
      properties:
        doctor_id: { type: integer }
        date: { type: string }
        time: { type: string }

    TreatmentInput:
      type: object
      required: [diagnosis, prescription, notes]
      properties:
        diagnosis: { type: string }
        prescription: { type: string }
        notes: { type: string }


paths:

  /api/register:
    post:
      summary: Register new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        "200":
          description: Patient registered
        "400":
          description: User already exists

  /api/login:
    post:
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginOutput'
        "400":
          description: Invalid credentials


  /api/me:
    get:
      summary: Get current logged user details
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Token invalid or expired


  /api/admin/doctors:
    post:
      summary: Admin adds a new doctor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorInput'
      responses:
        "200": { description: Doctor added }
        "403": { description: Admin access required }

    get:
      summary: Admin lists all doctors
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "403": { description: Admin access required }


  /api/patient/appointments:
    post:
      summary: Patient books appointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        "200":
          description: Appointment booked
        "403":
          description: Patient access required


  /api/doctor/appointments/{id}:
    put:
      summary: Doctor updates appointment with treatment notes
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreatmentInput'
      responses:
        "200":
          description: Treatment updated
        "403":
          description: Doctor access required
        "404":
          description: Appointment not found